back_end/WhereTo/requirements.txt
Django==3.2.7
polyline==1.4.0
pqdict==1.1.1
google-auth==2.1.0
